refactor : move js to src and utility to utils
