+++ bin/fibonacci/index.js
@@ -1,33 +1,43 @@
-import { generateFibonacci } from "../../pages/functional/js/fibonacci.js";
-import { Fibonacci } from "../../pages/oop/js/Fibonacci.js";
+#!/usr/bin/env node
 
-import { handleArgument } from "./handleArgument.js";
+import readline from "readline";
+import { generateFibonacci } from "../../src/functional/fibonacci.js";
+import { Fibonacci } from "../../src/oop/Fibonacci.js";
+import { validateApproach } from "../../src/utils/validateApproach.js";
+import { validateFibonacciMethod } from "../../src/utils/validateMethod.js";
+import { validateInputNumber } from "../../src/utils/validateInput.js";
 
-const cliArguments = process.argv.slice(2);
-const { type, method, sequence } = handleArgument(cliArguments);
-let result = 0;
+const rl = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout,
+});
 
-if (!["oop", "functional"].includes(type)) {
-  throw new Error(`invalid type ${type}. must be "oop" or "functional"`);
-}
+rl.question("Select the fibonacci approach (functional / oop) : ", approach => {
+  validateApproach(approach);
 
-if (!["loop", "recursive"].includes(method)) {
-  throw new Error(`invalid method ${method}. must be "loop" or "recursive"`);
-}
+  rl.question("Select the fibonacci method (loop / recursive) : ", method => {
+    validateFibonacciMethod(method);
 
-if (typeof parseInt(sequence) != "number") {
-  throw new Error(`invalid number ${sequence}. n should be a number"`);
-}
+    rl.question("Input the sequence : ", sequence => {
+      validateInputNumber(sequence);
 
-if (type == "functional") {
-  result = generateFibonacci(sequence, method);
-}
+      let result = [];
 
-if (type == "oop") {
-  const fibonacci = new Fibonacci(sequence);
-  result = fibonacci.generate(method);
-}
+      if (approach == "functional") {
+        result = generateFibonacci(sequence, method);
+      }
 
-console.log(`------generate fibonacci using ${method} (${type})------`);
-console.log(`fibonacci array with ${sequence} sequence :`);
-console.log(result);
+      if (approach == "oop") {
+        result = new Fibonacci(sequence).generate(method);
+      }
+
+      console.log(
+        `------generate fibonacci using ${method} (${approach})------`,
+      );
+      console.log(`fibonacci array with ${sequence} sequence : `);
+      console.log(result);
+
+      rl.close();
+    });
+  });
+});
