--- bin/shared/utils.js
@@ -1,3 +1,0 @@
-export const destructureArgument = argument => argument.split("=");
-
-export const validateArgument = argument => /^--\w+=\w+$/.test(argument);
