--- pages/functional/adapter/fibonacci.js
@@ -1,19 +1,0 @@
-import { generateFibonacci } from "../js/fibonacci.js";
-
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const sequence = event.target["sequence"].value;
-    const method = event.target["method"].value;
-
-    const numbers = generateFibonacci(sequence, method);
-
-    document.getElementById("result").innerHTML = numbers
-      .map(number => `<div class="result-item-fibonacci">${number}</div>`)
-      .join("");
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
