--- pages/oop/adapter/factorial.js
@@ -1,17 +1,0 @@
-import { Factorial } from "../js/Factorial.js";
-
-document.getElementById("form").addEventListener("submit", function (event) {
-  event.preventDefault();
-
-  try {
-    const n = event.target["n"].value;
-    const method = event.target["method"].value;
-
-    const result = new Factorial(n).count(method);
-
-    document.getElementById("result").textContent = result;
-  } catch (error) {
-    alert(error.message);
-    console.error(error);
-  }
-});
