+++ pages/oop/js/fizz-buzz.js
@@ -1,0 +1,27 @@
+import { FizzBuzz } from "../../../src/oop/FizzBuzz.js";
+
+document.getElementById("form").addEventListener("submit", function (event) {
+  event.preventDefault();
+
+  try {
+    const sequence = event.target["sequence"].value;
+
+    const fizzBuzzs = new FizzBuzz(sequence).generate();
+
+    document.getElementById("result").innerHTML = fizzBuzzs
+      .map(
+        (fizzBuzz, index) =>
+          `<div class="result-item-fizzbuzz bg-${
+            {
+              "fizz buzz": "red",
+              fizz: "blue",
+              buzz: "green",
+            }[fizzBuzz]
+          }">${index + 1} ${fizzBuzz}</div>`,
+      )
+      .join("");
+  } catch (error) {
+    alert(error.message);
+    console.error(error);
+  }
+});
